Removed more unneeded files.
