+++ includes/core/classes/class-block.php
@@ -109,7 +109,7 @@ class Block {
 		Blocks\Rsvp_Form::get_instance();
 		Blocks\Rsvp_Response::get_instance();
 		Blocks\Rsvp_Template::get_instance();
-		Blocks\Venue_Block::get_instance();
+		Blocks\Venue::get_instance();
 	}
 
 	/**
