+++ package.json
@@ -64,8 +64,6 @@
 		"moment-timezone": "^0.6.0",
 		"playwright": "^1.58.2",
 		"prettier": "^3.8.1",
-		"react-modal": "^3.16.3",
-		"react-tooltip": "^5.30.0",
 		"rimraf": "^6.1.3",
 		"stylelint-config-recommended-scss": "^17.0.0",
 		"stylelint-config-standard": "^40.0.0",
