+++ sonar-project.properties
@@ -15,7 +15,7 @@ sonar.test.inclusions=test/unit/php/**/*.php,test/unit/js/**/*.test.js
 
 # Exclusions
 sonar.exclusions=test/**/*,build/**/*,**/*.css,**/*.scss,webpack.config.js,package.json,includes/data/**/*
-sonar.coverage.exclusions=src/blocks/**/render.php,**/*.asset.php,includes/templates/**/*,includes/data/**/*,includes/core/classes/class-autoloader.php,includes/core/classes/traits/class-singleton.php,includes/core/requirements-check.php,includes/core/duplicate-check.php,gatherpress.php,src/index.js,src/editor.js,src/profile/index.js,src/settings/index.js,src/blocks/*/index.js,src/blocks/*/edit.js,src/blocks/*/view.js,src/blocks/*/save.js,src/blocks/*/template.js,src/blocks/*/templates/*.js,src/blocks/*/panels/**/*,src/blocks/*/types/**/*,src/blocks/*/fields/index.js,src/blocks/*/hooks/index.js,src/blocks/*/hooks/use-*.js,src/panels/**/*,src/modals/**/*,src/variations/**/*,src/components/**/*,src/supports/**/*,src/stores/index.js,src/helpers/urlrewrite.js,src/helpers/interactivity.js,src/blocks/events-list/events-list.js,src/blocks/online-event/online-event.js,src/blocks/venue/venue.js,src/blocks/rsvp-response/rsvp-manager.js
+sonar.coverage.exclusions=src/blocks/**/render.php,**/*.asset.php,includes/templates/**/*,includes/data/**/*,includes/core/classes/class-autoloader.php,includes/core/classes/traits/class-singleton.php,includes/core/requirements-check.php,includes/core/duplicate-check.php,gatherpress.php,src/index.js,src/editor.js,src/profile/index.js,src/settings/index.js,src/blocks/*/index.js,src/blocks/*/edit.js,src/blocks/*/view.js,src/blocks/*/save.js,src/blocks/*/template.js,src/blocks/*/templates/*.js,src/blocks/*/panels/**/*,src/blocks/*/types/**/*,src/blocks/*/fields/index.js,src/blocks/*/hooks/index.js,src/blocks/*/hooks/use-*.js,src/panels/**/*,src/modals/**/*,src/variations/**/*,src/components/**/*,src/supports/**/*,src/stores/index.js,src/helpers/urlrewrite.js,src/helpers/interactivity.js,src/blocks/rsvp-response/rsvp-manager.js
 sonar.cpd.exclusions=src/blocks/*/template.js,src/blocks/*/templates/**
 
 # Reports
