--- src/blocks/events-list/block.json
@@ -1,57 +1,0 @@
-{
-	"$schema": "https://schemas.wp.org/trunk/block.json",
-	"apiVersion": 2,
-	"name": "gatherpress/events-list",
-	"version": "1.0.1",
-	"title": "Events List",
-	"category": "gatherpress",
-	"icon": "list-view",
-	"example": {},
-	"description": "Displays events that are either upcoming or have occurred in the past.",
-	"attributes": {
-		"eventOptions": {
-			"type": "object",
-			"default": {
-				"descriptionLimit": 55,
-				"imageSize": "default",
-				"showRsvpResponse": true,
-				"showFeaturedImage": true,
-				"showDescription": true,
-				"showRsvp": true,
-				"showVenue": true
-			}
-		},
-		"datetimeFormat": {
-			"type": "string",
-			"default": "D, M j, Y, g:i a T"
-		},
-		"maxNumberOfEvents": {
-			"type": "integer",
-			"default": 5
-		},
-		"topics": {
-			"type": "array",
-			"items": {
-				"type": "object"
-			}
-		},
-		"venues": {
-			"type": "array",
-			"items": {
-				"type": "object"
-			}
-		},
-		"type": {
-			"type": "string",
-			"default": "upcoming"
-		}
-	},
-	"supports": {
-		"html": false
-	},
-	"textdomain": "gatherpress",
-	"editorScript": "file:./index.js",
-	"style": "file:./style-index.css",
-	"viewScript": "file:./events-list.js",
-	"render": "file:./render.php"
-}
