--- src/blocks/events-list/style.scss
@@ -1,219 +1,0 @@
-@use "../buttons";
-@use "../mixins";
-@use "../modal";
-@use "../status";
-
-.gatherpress-events-list {
-	&__datetime {
-		font-size: var(--wp--preset--font-size--small);
-		margin-bottom: 0.5rem;
-		padding: 0;
-		text-transform: uppercase;
-	}
-
-	&__excerpt {
-		margin-bottom: 1rem;
-
-		p {
-			margin: 0;
-			padding: 0;
-		}
-	}
-
-	.gatherpress-events-list__footer {
-		border-bottom: 1px solid #ccc;
-		display: flex;
-		margin-bottom: 2rem;
-		padding-bottom: 2rem;
-
-		@include mixins.for-core-mobile {
-			line-height: 2.75rem;
-		}
-
-		.gatherpress-rsvp {
-			margin-bottom: 0;
-
-			@include mixins.for-core-mobile {
-				flex-direction: column;
-			}
-		}
-
-		.gatherpress-rsvp-response {
-			&__items {
-				margin-left: 1rem;
-				flex-grow: 2;
-				grid-template-columns: repeat(auto-fill, minmax(min(1rem, 100%), 1fr));
-			}
-
-			&__member-avatar {
-				display: flex;
-				align-items: center;
-				margin-bottom: 0;
-				margin-left: -1rem;
-				width: 3.125rem;
-
-				img {
-					width: 3.125rem;
-				}
-			}
-		}
-	}
-
-	&__header {
-		display: flex;
-		gap: 2rem;
-		margin-bottom: 1rem;
-	}
-
-	&__image {
-		float: right;
-		flex-shrink: 0;
-		margin: 0 0 1rem 1rem;
-
-		a {
-			display: block;
-		}
-	}
-
-	&__info {
-		flex-grow: 1;
-		padding: 0;
-	}
-
-	&__title {
-		font-size: var(--wp--preset--font-size--large);
-		margin-bottom: 1rem;
-		padding: 0;
-	}
-
-	&__venue {
-		align-items: center;
-		display: flex;
-		gap: 0.5rem;
-		margin-bottom: 1rem;
-	}
-
-	// Old RSVP Styles.
-	.gatherpress-modal__rsvp {
-		.gatherpress-modal {
-			&__header {
-				font-size: var(--wp--preset--font-size--medium);
-				font-weight: 600;
-			}
-
-			&__guests {
-				display: flex;
-				gap: 0.5rem;
-				margin-bottom: 0.5rem;
-			}
-
-			&__anonymous {
-				display: flex;
-				gap: 0.5rem;
-				margin-bottom: 0.5rem;
-			}
-		}
-	}
-
-	.gatherpress-rsvp {
-		align-items: center;
-		display: flex;
-		column-gap: 1rem;
-		justify-content: center;
-		margin-bottom: 1rem;
-	}
-
-	// Old RSVP Response Styles.
-	.gatherpress-rsvp-response {
-		&__anchor {
-			color: #000;
-			display: block;
-			padding: 0.5rem;
-			white-space: nowrap;
-		}
-
-		&__header {
-			align-items: center;
-			display: flex;
-			gap: 0.75rem;
-			margin-bottom: 1rem;
-		}
-
-		&__items {
-			display: grid;
-			gap: 1rem;
-			grid-template-columns: repeat(4, minmax(0, 1fr));
-
-			@include mixins.for-core-mobile {
-				grid-template-columns: repeat(2, minmax(0, 1fr));
-			}
-		}
-
-		&__item {
-			background-color: #fff;
-			border-radius: 0.25rem;
-			box-shadow: 0 0 0.25rem 0 #ccc;
-			gap: 1rem;
-			padding: 1rem;
-			text-align: center;
-
-			& a {
-				display: block;
-				overflow: hidden;
-				text-overflow: ellipsis;
-				white-space: nowrap;
-			}
-		}
-
-		&__member-avatar {
-			margin: 0;
-
-			& img {
-				border-radius: 9999px;
-			}
-		}
-
-		&__member-role {
-			color: #646970;
-			font-size: var(--wp--preset--font-size--small);
-		}
-
-		&__navigation {
-			border: 1px solid #ccc;
-			border-radius: 0.4rem;
-			flex-direction: column;
-			overflow: hidden;
-			position: absolute;
-			width: 15rem;
-		}
-
-		&__navigation-item {
-			align-items: center;
-			background-color: #fff;
-			border-top: 1px solid #ccc;
-			display: flex;
-			margin-top: -1px;
-		}
-
-		&__no-responses {
-			grid-column: 1 / -1;
-		}
-
-		&__see-all {
-			margin-left: auto;
-		}
-
-		&__title {
-			font-size: var(--wp--preset--font-size--medium);
-			font-weight: 600;
-			margin-bottom: 1rem;
-		}
-
-		// UI fix in editor.
-		.edit-post-visual-editor & {
-			.components-button.has-icon {
-				padding: 0 2px;
-			}
-		}
-	}
-}
