+++ src/blocks/venue-map/block.json
@@ -31,7 +31,8 @@
 	},
 	"textdomain": "gatherpress",
 	"editorScript": "file:./index.js",
+	"editorStyle": "file:./index.css",
+	"viewScript": "file:./view.js",
 	"style": "file:./style-index.css",
-	"viewScript": "file:../venue/venue.js",
 	"render": "file:./render.php"
 }
