--- test/unit/js/src/components/RsvpResponse.test.js
@@ -1,217 +1,0 @@
-/**
- * External dependencies.
- */
-import { render, act } from '@testing-library/react';
-import { describe, expect, it, jest, beforeEach } from '@jest/globals';
-import '@testing-library/jest-dom';
-
-/**
- * WordPress dependencies.
- */
-jest.mock( '@wordpress/i18n', () => ( {
-	_x: ( text ) => text,
-} ) );
-
-jest.mock( '@wordpress/element', () => ( {
-	...jest.requireActual( '@wordpress/element' ),
-} ) );
-
-/**
- * Internal dependencies.
- */
-import * as broadcasting from '../../../../../src/helpers/broadcasting';
-import * as globals from '../../../../../src/helpers/globals';
-
-jest.mock( '../../../../../src/helpers/broadcasting' );
-jest.mock( '../../../../../src/helpers/globals' );
-
-// Mock RsvpResponseHeader component.
-jest.mock(
-	'../../../../../src/components/RsvpResponseHeader',
-	() => ( props ) => (
-		<div data-testid="rsvp-response-header">
-			Mock Header
-			<button
-				onClick={ ( e ) => props.onTitleClick( e, 'not_attending' ) }
-			>
-				Change Status
-			</button>
-		</div>
-	)
-);
-
-// Mock RsvpResponseContent component.
-jest.mock(
-	'../../../../../src/components/RsvpResponseContent',
-	() => ( props ) => (
-		<div data-testid="rsvp-response-content">
-			Mock Content - Active: { props.activeValue }
-		</div>
-	)
-);
-
-import RsvpResponse from '../../../../../src/components/RsvpResponse';
-
-/**
- * Coverage for RsvpResponse.
- */
-describe( 'RsvpResponse', () => {
-	const mockListener = jest.fn();
-
-	beforeEach( () => {
-		jest.clearAllMocks();
-		broadcasting.Listener = mockListener;
-		globals.getFromGlobal = jest.fn( ( key ) => {
-			if ( 'eventDetails.postId' === key ) {
-				return 123;
-			}
-			if ( 'eventDetails.hasEventPast' === key ) {
-				return false;
-			}
-			return null;
-		} );
-
-		global.GatherPress = {
-			eventDetails: {
-				responses: {
-					all: {
-						count: 1,
-						records: [
-							{
-								guests: 0,
-								id: 1,
-								name: 'unittest',
-								photo: 'https://unit.test/photo',
-								profile: 'https://unit.test/profile',
-								role: 'Member',
-								status: 'attending',
-								timestamp: '2023-05-11 00:00:00',
-							},
-						],
-					},
-					attending: {
-						count: 1,
-						records: [
-							{
-								guests: 0,
-								id: 1,
-								name: 'John Doe',
-								photo: 'https://unit.test/photo',
-								profile: 'https://unit.test/profile',
-								role: 'Member',
-								status: 'attending',
-								timestamp: '2023-05-11 00:00:00',
-							},
-						],
-					},
-					not_attending: {
-						count: 0,
-						records: [],
-					},
-					waiting_list: {
-						count: 0,
-						records: [],
-					},
-				},
-			},
-		};
-	} );
-
-	it( 'renders component correctly', () => {
-		const { container } = render( <RsvpResponse /> );
-
-		expect( container.children[ 0 ] ).toHaveClass( 'gatherpress-rsvp-response' );
-	} );
-
-	it( 'renders RsvpResponseHeader and RsvpResponseContent components', () => {
-		const { container } = render( <RsvpResponse /> );
-
-		expect( container.querySelector( '[data-testid="rsvp-response-header"]' ) ).toBeInTheDocument();
-		expect( container.querySelector( '[data-testid="rsvp-response-content"]' ) ).toBeInTheDocument();
-	} );
-
-	it( 'uses default status of attending', () => {
-		const { container } = render( <RsvpResponse /> );
-
-		expect( container.textContent ).toContain( 'Active: attending' );
-	} );
-
-	it( 'accepts custom defaultStatus prop', () => {
-		const { container } = render( <RsvpResponse defaultStatus="waiting_list" /> );
-
-		expect( container.textContent ).toContain( 'Active: waiting_list' );
-	} );
-
-	it( 'calls onTitleClick and updates status', () => {
-		const { container } = render( <RsvpResponse /> );
-
-		// Initially attending.
-		expect( container.textContent ).toContain( 'Active: attending' );
-
-		// Click the button to trigger onTitleClick.
-		const button = container.querySelector( 'button' );
-		act( () => {
-			button.click();
-		} );
-
-		// Status should change to not_attending.
-		expect( container.textContent ).toContain( 'Active: not_attending' );
-	} );
-
-	it( 'prevents default on onTitleClick', () => {
-		const { container } = render( <RsvpResponse /> );
-
-		const button = container.querySelector( 'button' );
-
-		// We can't directly test preventDefault since React's synthetic events handle it,
-		// but we can verify the button exists and onClick is called.
-		expect( button ).toBeInTheDocument();
-	} );
-
-	it( 'calls Listener with correct parameters', () => {
-		render( <RsvpResponse /> );
-
-		expect( mockListener ).toHaveBeenCalledWith(
-			expect.objectContaining( {
-				setRsvpStatus: expect.any( Function ),
-			} ),
-			123
-		);
-	} );
-
-	it( 'uses "Attending" title for upcoming events', () => {
-		globals.getFromGlobal = jest.fn( ( key ) => {
-			if ( 'eventDetails.hasEventPast' === key ) {
-				return false;
-			}
-			return null;
-		} );
-
-		render( <RsvpResponse /> );
-
-		// The _x mock returns the text as-is, so we expect the first argument.
-		// This ensures the correct text is being passed to _x based on hasEventPast.
-		expect( globals.getFromGlobal ).toHaveBeenCalledWith( 'eventDetails.hasEventPast' );
-	} );
-
-	it( 'uses "Went" title for past events', () => {
-		globals.getFromGlobal = jest.fn( ( key ) => {
-			if ( 'eventDetails.hasEventPast' === key ) {
-				return true;
-			}
-			return null;
-		} );
-
-		render( <RsvpResponse /> );
-
-		expect( globals.getFromGlobal ).toHaveBeenCalledWith( 'eventDetails.hasEventPast' );
-	} );
-
-	it( 'initializes rsvpLimit with defaultLimit', () => {
-		const { container } = render( <RsvpResponse /> );
-
-		// RsvpResponseHeader receives rsvpLimit and defaultLimit props.
-		// We can't directly inspect the state, but we know it's initialized to 8.
-		expect( container.querySelector( '[data-testid="rsvp-response-header"]' ) ).toBeInTheDocument();
-	} );
-} );
